requirements.txt
customtkinter
SpeechRecognition
gTTS
pygame
pyttsx3
pyaudio
thefuzz
groq
yt-dlp
python-dotenv
